test: Update profiling sample
